+++ src/Phan/Language/EmptyUnionType.php
@@ -1562,4 +1562,9 @@ final class EmptyUnionType extends UnionType
     {
         return false;
     }
+
+    public function convertUndefinedToNullable() : UnionType
+    {
+        return $this;
+    }
 }